inline concatenate & change parameters

+++ src/ILS.h
@@ -1,6 +1,6 @@
 #ifndef ILS_H
 #define ILS_H
 
-int ILS(double **matrizAdj, int dimensao);
+float ILS(double **matrizAdj, int dimensao);
 
 #endif
